refactore: 응답자 결과 화면 서버 컴포넌트로 수정
